work in progress

+++ stage14_floating_point/integration_tests/floating_point/test_basic_dbl__40.c
@@ -1,6 +1,6 @@
 int main() {
-    float a = 1.5f, b = 2.5f;
-    float c = a + b;
+    double a = 1.5f, b = 2.5f;
+    double c = a + b;
     return (int)(c + 36.0f);   // 40
 }
